Restructure of hermes package
