+++ Hermes/appscale/hermes/constants.py
@@ -1,69 +1,41 @@
-""" Constants that are used for Hermes functionality (backup/restore/etc.). """
+# Path to haproxy stats sockets
+HAPROXY_APPS_STATS_SOCKET_PATH = '/etc/haproxy/stats'
+HAPROXY_SERVICES_STATS_SOCKET_PATH = '/etc/haproxy/service-stats'
+# Path to haproxy stats sockets
+HAPROXY_APPS_CONFIGS_DIR = '/etc/haproxy/app-sites-enabled/'
+HAPROXY_SERVICES_CONFIGS_DIR = '/etc/haproxy/service-sites-enabled/'
 
-# The user account type for deploying the appscalesensor app.
+# Path to dictionary to write profile log
+PROFILE_LOG_DIR = '/var/log/appscale/profile'
 
-ACCOUNT_TYPE = 'user'
+# The amount of time to wait for local stats from a slave node.
+STATS_REQUEST_TIMEOUT = 60
 
-# Location where deployed app source code resides in an AppScale deployment.
-APP_DIR_LOCATION = '/opt/appscale/apps'
+# Stats which were produce less than X seconds ago is considered as current
+ACCEPTABLE_STATS_AGE = 10
 
-# The app id for the appscalesensor application.
-APPSCALE_SENSOR = 'appscalesensor'
+# The ZooKeeper location for storing Hermes configurations
+NODES_STATS_CONFIGS_NODE = '/appscale/stats/profiling/nodes'
+PROCESSES_STATS_CONFIGS_NODE = '/appscale/stats/profiling/processes'
+PROXIES_STATS_CONFIGS_NODE = '/appscale/stats/profiling/proxies'
 
-# The port br_service listens to.
-BR_SERVICE_PORT = 8423
 
-# The br_service path for starting a new task.
-BR_SERVICE_PATH = '/'
+class _MissedValue(object):
+  """
+  Instance of this private class denotes missed value.
+  It's used to denote values of stats properties which are missed
+  in haproxy stats.
+  """
 
-# The suffix for backup files from a DB master node.
-DB_MASTER_OBJECT_NAME = '/cassandra/db_master.tar.gz'
+  def __nonzero__(self):
+    return False
 
-# Enable DEBUG logging.
-DEBUG = False
+  def __repr__(self):
+    return ''
 
-# The suffix for backup files from a DB slave node.
-# The slave ID is populated at runtime.
-DB_SLAVE_OBJECT_NAME = '/cassandra/db_slave_{0}.tar.gz'
 
-# The file suffix to match for the tarred file to be deployed.
-FILE_SUFFIX = 'tar.gz'
+MISSED = _MissedValue()
 
-# The port Hermes listens to.
-HERMES_PORT = 4378
-
-# A constant representing the Tornado HTTPError.
-HTTPError = 'HTTPError'
-
-# The interval between polls for new tasks.
-POLLING_INTERVAL = 30*1000    # 30 seconds.
-
-# The AppScale Portal URL for getting new tasks.
-PORTAL_URL = 'https://portal.appscale.com'
-
-# The AppScale Portal path for getting new tasks.
-PORTAL_POLL_PATH = '/get_appscale_task'
-
-# The AppScale Portal path for sending all stats.
-PORTAL_STATS_PATH = '/deployments/{}/all_stats'
-
-# The AppScale Portal path for reporting task status.
-PORTAL_STATUS_PATH = '/report_appscale_task'
-
-# A list of required parameters that define a task.
-REQUIRED_KEYS = ['task_id', 'type', 'bucket_name', 'storage']
-
-# The amount of time to wait for a node backup in seconds.
-REQUEST_TIMEOUT = 12*60*60
-
-# The list of supported tasks.
-SUPPORTED_TASKS = ['backup', 'restore']
-
-# The interval for checking for registered deployments.
-UPLOAD_SENSOR_INTERVAL = 60*1000    # 60 seconds.
-
-# The username for deploying the appscalesensor app.
-USER_EMAIL = 'appscale_user@appscale.local'
 
 # Name of header where secret should be passed
 SECRET_HEADER = 'Appscale-Secret'
@@ -77,8 +49,3 @@ class HTTP_Codes(object):
   HTTP_INTERNAL_ERROR = 500
   HTTP_NOT_IMPLEMENTED = 501
 
-
-class TaskTypes(object):
-  """ A class containing supported task types. """
-  CASSANDRA_BACKUP = 'cassandra_backup'
-  CASSANDRA_RESTORE = 'cassandra_restore'
